Addressed some of the requested changes.
